backend/lambdas/data_collector/requirements.txt
requests>=2.28.0
psycopg2-binary>=2.9.5
boto3>=1.26.0
